fix: Claims for the same workload are created repeatedly

+++ pkg/controller/ippool_controller.go
@@ -56,7 +56,12 @@ func (r *IPPoolReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctr
 		return ctrl.Result{}, client.IgnoreNotFound(err)
 	}
 
-	if !rp.DeletionTimestamp.IsZero() {
+	var riList requeueipv1.IPList
+	if err := r.client.List(ctx, &riList, client.MatchingLabels{consts.LabelRefIPPool: rp.Name}); err != nil {
+		return ctrl.Result{}, err
+	}
+
+	if len(riList.Items) == 0 && !rp.DeletionTimestamp.IsZero() {
 		if err := r.client.DeleteAllOf(
 			ctx,
 			&requeueipv1.IPBlock{},
@@ -70,11 +75,6 @@ func (r *IPPoolReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctr
 
 		controllerutil.RemoveFinalizer(&rp, consts.RFinalizer)
 		return ctrl.Result{}, r.client.Update(ctx, &rp)
-	}
-
-	var riList requeueipv1.IPList
-	if err := r.client.List(ctx, &riList, client.MatchingLabels{consts.LabelRefIPPool: rp.Name}); err != nil {
-		return ctrl.Result{}, err
 	}
 
 	// Calculate the total, used, and available range of the IPPool.
